// cat-launcher/src/generated-types/ColorTheme.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Represents a color theme for the game.
 */
export type ColorTheme = { 
/**
 * A unique identifier for the theme.
 */
id: string, 
/**
 * The display name of the theme.
 */
name: string, 
/**
 * The absolute path to the theme's configuration file.
 */
path: string, };