// cat-launcher/src/generated-types/GameEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GameErrorPayload } from "./GameErrorPayload";
import type { GameExitPayload } from "./GameExitPayload";

/**
 * Events emitted during the game session.
 */
export type GameEvent = { "type": "Log", "payload": string } | { "type": "Exit", "payload": GameExitPayload } | { "type": "Error", "payload": GameErrorPayload };