// cat-launcher/src/generated-types/GameVariantInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GameVariant } from "./GameVariant";

/**
 * Information about a game variant, suitable for display in the UI.
 */
export type GameVariantInfo = { 
/**
 * The unique identifier for the game variant.
 */
id: GameVariant, 
/**
 * The human-readable name of the game variant.
 */
name: string, };